+++ dsl/common/cloud/dsl-common-aws/src/main/kotlin/io/kotless/dsl/cloud/aws/Permissions.kt
@@ -25,3 +25,11 @@ annotation class Cognito(val userPoolsId: String, val level: PermissionLevel)
 /** Delegates permissions to specified SecretManager records by pattern with annotation */
 @Target(AnnotationTarget.FUNCTION, AnnotationTarget.CLASS, AnnotationTarget.PROPERTY)
 annotation class SecretManager(val pattern: String, val level: PermissionLevel)
+
+/** Delegates permissions to specified SNS topic to entity with annotation */
+@Target(AnnotationTarget.FUNCTION, AnnotationTarget.CLASS, AnnotationTarget.PROPERTY)
+annotation class SNSTopic(val topicArn: String, val level: PermissionLevel, val region: String = "")
+
+/** Marks a function to consume messages from an SNS topic */
+@Target(AnnotationTarget.FUNCTION)
+annotation class SNSEvent(val topicName: String, val region: String = "")