+++ CHANGELOG.md
@@ -4,8 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 # 0.3.4 - 2025-12-26
 
+* Breaking change: you should remove dependency in "kotless-lang-aws" as its no longer part of the project
 * removed ktor & kotless dsl
 * removed azure support
+* support sns consumption
 
 # 0.3.3 - 2023-12-25